총기 클래스 추가 및 탄환 데미지에 거너 2-2특성 적용
