+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/abilities/gunner/Buckshot.java
@@ -1,0 +1,6 @@
+package com.shatteredpixel.shatteredpixeldungeon.actors.hero.abilities.gunner;
+
+import com.shatteredpixel.shatteredpixeldungeon.actors.hero.abilities.ArmorAbility;
+
+public class Buckshot  extends ArmorAbility {
+}
