+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/abilities/gunner/GrenadeLauncher.java
@@ -1,0 +1,6 @@
+package com.shatteredpixel.shatteredpixeldungeon.actors.hero.abilities.gunner;
+
+import com.shatteredpixel.shatteredpixeldungeon.actors.hero.abilities.ArmorAbility;
+
+public class GrenadeLauncher  extends ArmorAbility {
+}
