+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/gun/GunsWeapon.java
@@ -1,0 +1,6 @@
+package com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.gun;
+
+import com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.MeleeWeapon;
+
+public class GunsWeapon extends MeleeWeapon {
+}
