+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/gun/Pistol.java
@@ -19,58 +19,23 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  */
 
-package com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee;
+package com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.gun;
 
 import com.shatteredpixel.shatteredpixeldungeon.Assets;
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Actor;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
-import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Momentum;
-import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.RevealedArea;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
-import com.shatteredpixel.shatteredpixeldungeon.actors.hero.HeroSubClass;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Talent;
 import com.shatteredpixel.shatteredpixeldungeon.effects.CellEmitter;
-import com.shatteredpixel.shatteredpixeldungeon.effects.Splash;
 import com.shatteredpixel.shatteredpixeldungeon.effects.particles.BlastParticle;
 import com.shatteredpixel.shatteredpixeldungeon.effects.particles.SmokeParticle;
-import com.shatteredpixel.shatteredpixeldungeon.items.ArcaneResin;
-import com.shatteredpixel.shatteredpixeldungeon.items.bombs.Firebomb;
-import com.shatteredpixel.shatteredpixeldungeon.items.bombs.Flashbang;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.AlchemicalCatalyst;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.Potion;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfExperience;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfFrost;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfHaste;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfHealing;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfInvisibility;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfLevitation;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfLiquidFlame;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfMindVision;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfParalyticGas;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfPurity;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfToxicGas;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.brews.BlizzardBrew;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.brews.CausticBrew;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.brews.InfernalBrew;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.brews.ShockingBrew;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfCleansing;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfCorrosiveGas;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfDragonsBreath;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfEarthenArmor;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfHolyFuror;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfMagicalSight;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfShielding;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfShroudingFog;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfSnapFreeze;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfStamina;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.exotic.PotionOfStormClouds;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfReload;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfSharpshooting;
-import com.shatteredpixel.shatteredpixeldungeon.items.spells.ArcaneCatalyst;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.SpiritBow;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.Weapon;
+import com.shatteredpixel.shatteredpixeldungeon.items.weapon.melee.MeleeWeapon;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.missiles.MissileWeapon;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.scenes.CellSelector;
@@ -83,7 +48,7 @@ import com.watabou.utils.Random;
 
 import java.util.ArrayList;
 
-public class DualPistol extends MeleeWeapon {
+public class Pistol extends GunsWeapon {
 
     public static final String AC_SHOOT		= "SHOOT";
     public static final String AC_RELOAD = "RELOAD";
@@ -98,7 +63,7 @@ public class DualPistol extends MeleeWeapon {
         defaultAction = AC_SHOOT;
         usesTargeting = true;
 
-        image = ItemSpriteSheet.DUAL_PISTOL;                                  //if you make something different guns, you should change this
+        image = ItemSpriteSheet.PISTOL;                                  //if you make something different guns, you should change this
         hitSound = Assets.Sounds.HIT_CRUSH;
         hitSoundPitch = 0.8f;
 
@@ -162,7 +127,7 @@ public class DualPistol extends MeleeWeapon {
             }
         }
         if (action.equals(AC_RELOAD)) {
-            max_round = 8;                                                                  //if you make something different guns, you should change this
+            max_round = 4;                                                                  //if you make something different guns, you should change this
             if (round == max_round){
                 GLog.w(Messages.get(this, "already_loaded"));
             }
@@ -171,7 +136,7 @@ public class DualPistol extends MeleeWeapon {
     }
 
     public void reload() {
-        max_round = 8;                                                                      //if you make something different guns, you should change this
+        max_round = 4;                                                                      //if you make something different guns, you should change this
         curUser.spend(reload_time);
         curUser.busy();
         Sample.INSTANCE.play(Assets.Sounds.UNLOCK, 2, 1.1f);
@@ -188,7 +153,7 @@ public class DualPistol extends MeleeWeapon {
 
     @Override
     public String status() {
-        max_round = 8;                                                                      //if you make something different guns, you should change this
+        max_round = 4;                                                                      //if you make something different guns, you should change this
         return Messages.format(TXT_STATUS, round, max_round);
     }
 
@@ -199,30 +164,30 @@ public class DualPistol extends MeleeWeapon {
 
     public int min(int lvl) {
         return tier +                                                                      //if you make something different guns, you should change this
-               lvl;                                                                        //if you make something different guns, you should change this
+                lvl;                                                                        //if you make something different guns, you should change this
     }
 
     public int max(int lvl) {
         return 3 * (tier + 1) +                                                            //if you make something different guns, you should change this
-               lvl;                                                           //if you make something different guns, you should change this
+                lvl;                                                           //if you make something different guns, you should change this
     }
 
     public int Bulletmin(int lvl) {
-        return tier +                                                                  //if you make something different guns, you should change this
-               lvl +                                                                  //if you make something different guns, you should change this
-               RingOfSharpshooting.levelDamageBonus(Dungeon.hero);
+        return 2 * tier +                                                                  //if you make something different guns, you should change this
+                lvl      +                                                                  //if you make something different guns, you should change this
+                RingOfSharpshooting.levelDamageBonus(Dungeon.hero);
     }
 
     public int Bulletmax(int lvl) {
-        return 2 * (tier)   +                                                           //if you make something different guns, you should change this
-               lvl * (tier) +                                                           //if you make something different guns, you should change this
-               RingOfSharpshooting.levelDamageBonus(Dungeon.hero);
+        return 4 * (tier)   +                                                           //if you make something different guns, you should change this
+                lvl * (tier) +                                                           //if you make something different guns, you should change this
+                RingOfSharpshooting.levelDamageBonus(Dungeon.hero);
     }
 
     @Override
     public String info() {
 
-        max_round = 8;                                                                       //if you make something different guns, you should change this
+        max_round = 4;                                                                       //if you make something different guns, you should change this
         reload_time = 2f* RingOfReload.reloadMultiplier(Dungeon.hero);         //if you make something different guns, you should change this;
 
         String info = desc();
@@ -234,16 +199,16 @@ public class DualPistol extends MeleeWeapon {
             } else if (Dungeon.hero.STR() > STRReq()){
                 info += " " + Messages.get(Weapon.class, "excess_str", Dungeon.hero.STR() - STRReq());
             }
-            info += "\n\n" + Messages.get(DualPistol.class, "stats_known",
-                    Bulletmin(DualPistol.this.buffedLvl()),
-                    Bulletmax(DualPistol.this.buffedLvl()),
+            info += "\n\n" + Messages.get(Pistol.class, "stats_known",
+                    Bulletmin(Pistol.this.buffedLvl()),
+                    Bulletmax(Pistol.this.buffedLvl()),
                     round, max_round, reload_time);
         } else {
             info += "\n\n" + Messages.get(MeleeWeapon.class, "stats_unknown", tier, min(0), max(0), STRReq(0));
             if (STRReq(0) > Dungeon.hero.STR()) {
                 info += " " + Messages.get(MeleeWeapon.class, "probably_too_heavy");
             }
-            info += "\n\n" + Messages.get(DualPistol.class, "stats_unknown",
+            info += "\n\n" + Messages.get(Pistol.class, "stats_unknown",
                     Bulletmin(0),
                     Bulletmax(0),
                     round, max_round, reload_time);
@@ -304,13 +269,13 @@ public class DualPistol extends MeleeWeapon {
         return super.baseDelay(owner);
     }                   //공격 속도
 
-    public DualPistol.Bullet knockBullet(){
-        return new DualPistol.Bullet();
+    public Pistol.Bullet knockBullet(){
+        return new Pistol.Bullet();
     }
     public class Bullet extends MissileWeapon {
 
         {
-            image = ItemSpriteSheet.DUAL_BULLET;
+            image = ItemSpriteSheet.SINGLE_BULLET;
 
             hitSound = Assets.Sounds.PUFF;
             tier = 2;                                                                            //if you make something different guns, you should change this
@@ -318,50 +283,51 @@ public class DualPistol extends MeleeWeapon {
 
         @Override
         public int damageRoll(Char owner) {
+            int bulletdamage;
             if (owner.buff(Momentum.class) != null && owner.buff(Momentum.class).freerunning()) {
-                return Math.round(Random.NormalIntRange(Bulletmin(DualPistol.this.buffedLvl()), Bulletmax(DualPistol.this.buffedLvl())) * (1f + 0.15f * ((Hero) owner).pointsInTalent(Talent.PROJECTILE_MOMENTUM)));
+                bulletdamage = Math.round(Random.NormalIntRange(Bulletmin(Pistol.this.buffedLvl()), Bulletmax(Pistol.this.buffedLvl())) * (1f + 0.15f * ((Hero) owner).pointsInTalent(Talent.PROJECTILE_MOMENTUM)));
             } else {
-                return Random.NormalIntRange(Bulletmin(DualPistol.this.buffedLvl()), Bulletmax(DualPistol.this.buffedLvl()));
+                bulletdamage = Random.NormalIntRange(Bulletmin(Pistol.this.buffedLvl()), Bulletmax(Pistol.this.buffedLvl()));
             }
+            bulletdamage *= (1 + 0.1f * Dungeon.hero.pointsInTalent(Talent.ARM_ENHANCE));
+            return bulletdamage;
         }
 
         @Override
         public boolean hasEnchant(Class<? extends Enchantment> type, Char owner) {
-            return DualPistol.this.hasEnchant(type, owner);
+            return Pistol.this.hasEnchant(type, owner);
         }
 
         @Override
         public int proc(Char attacker, Char defender, int damage) {
-            return DualPistol.this.proc(attacker, defender, damage);
+            return Pistol.this.proc(attacker, defender, damage);
         }
 
         @Override
         public float delayFactor(Char user) {
-            return DualPistol.this.delayFactor(user);
+            return Pistol.this.delayFactor(user);
         }
 
         @Override
         public int STRReq(int lvl) {
-            return STRReq(tier, DualPistol.this.buffedLvl());
+            return STRReq(tier, Pistol.this.buffedLvl());
         }
 
         @Override
         protected void onThrow( int cell ) {
-            for (int i=1; i<=2; i++) {                                                           //i<=n에서 n이 반복하는 횟수, 즉 발사 횟수
-                Char enemy = Actor.findChar(cell);
-                if (enemy == null || enemy == curUser) {
-                    parent = null;
+            Char enemy = Actor.findChar( cell );
+            if (enemy == null || enemy == curUser) {
+                parent = null;
+                CellEmitter.get(cell).burst(SmokeParticle.FACTORY, 2);
+                CellEmitter.center(cell).burst(BlastParticle.FACTORY, 2);
+            } else {
+                if (!curUser.shoot( enemy, this )) {
                     CellEmitter.get(cell).burst(SmokeParticle.FACTORY, 2);
                     CellEmitter.center(cell).burst(BlastParticle.FACTORY, 2);
-                } else {
-                    if (!curUser.shoot(enemy, this)) {
-                        CellEmitter.get(cell).burst(SmokeParticle.FACTORY, 2);
-                        CellEmitter.center(cell).burst(BlastParticle.FACTORY, 2);
-                    }
                 }
-                round--;
-                updateQuickslot();
             }
+            round --;
+            updateQuickslot();
         }
 
         @Override
